ref(api): Split pagination logic into submodule
